:alien: Synchroniser la langue initiale
