+++ source/entry.dev.js
@@ -8,6 +8,9 @@ import debounceFormChangeActions from './debounceFormChangeActions'
 import computeThemeColours from './components/themeColours'
 import { getIframeOption, getUrl } from './utils'
 
+import { rules, rulesFr } from 'Engine/rules'
+import lang from './i18n'
+
 import App from './containers/App.dev'
 
 let initialStore = {
@@ -20,7 +23,8 @@ let enhancer = compose(
 	DevTools.instrument()
 )
 
-let store = createStore(reducers, initialStore, enhancer)
+let initialRules = lang == 'en' ? rules : rulesFr
+let store = createStore(reducers(initialRules), initialStore, enhancer)
 let anchor = document.querySelector('#js')
 
 render(<App store={store} />, anchor)
